+++ CHANGELOG.md
@@ -501,3 +501,4 @@
 * DamageResistanceNode (actor damage resistance) now uses custom property drawer
 * HitboxComponent now uses custom property drawer for hit location and material
 * Hit puffs and entity controllers now use custom property drawer for hit material
+* RpgDefaultValues is now a non-static class and implementes IRpgDefaultValues interface